Refactor Learnly workspace presenter into modular components
